Refactor sandbox file names
