+++ src/DonateTo.WebApplication/ClientApp/src/app/admin/organization/organization-create/organization-create.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnDestroy, ViewChild } from '@angular/core';
 import { Sandbox } from 'src/app/shared/sandbox/base.sandbox';
 import { OrganizationFormComponent } from '../organization-form/organization-form.component';
-import { OrganizationSandbox } from '../organization-sandbox';
+import { OrganizationSandbox } from '../organization.sandbox';
 import { Router } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { OrganizationModel } from 'src/app/shared/models';
