+++ src/DonateTo.WebApplication/ClientApp/src/app/admin/organization/organization-form/organization-step-address/organization-step-address.component.ts
@@ -2,7 +2,7 @@ import { Component, EventEmitter, Input, OnDestroy, OnInit, Output, ViewChild } 
 import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
 import { AddressModel, CityModel, ColumnItem, ContactModel, CountryModel, StateModel } from 'src/app/shared/models';
 import { Subscription } from 'rxjs';
-import { OrganizationSandbox } from '../../organization-sandbox';
+import { OrganizationSandbox } from '../../organization.sandbox';
 import { NzModalRef } from 'ng-zorro-antd';
 import { OrganizationStepContactComponent } from '../organization-step-contact/organization-step-contact.component';
 import { BranchModalComponent } from './branch-modal/branch-modal.component';
