+++ src/DonateTo.WebApplication/ClientApp/src/app/admin/organization/organization-form/organization-step-contact/organization-step-contact.component.ts
@@ -1,7 +1,7 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { ContactModel, OrganizationModel } from 'src/app/shared/models';
-import { OrganizationSandbox } from '../../organization-sandbox';
+import { OrganizationSandbox } from '../../organization.sandbox';
 
 @Component({
   selector: 'app-organization-step-contact',
